@@ -3,7 +3,7 @@ import { FormBuilder, Validators } from '@angular/forms';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
 import { map, startWith } from 'rxjs/operators';
-import { MatCardModule, MatDialogRef } from '@angular/material';
+import { MatDialogRef } from '@angular/material/dialog';
 import { ValidationMessages, UsernameValidator } from './validations/login.validations';
 
 import { Constants } from '../utilities/constants/Constants';
@@ -19,27 +19,44 @@ export class LoginRegisterComponent implements OnInit {
 
   @ViewChild("userFocus", { static: false }) userFocus: ElementRef;
 
-  loginForm;
-  forgotForm;
+  loginForm: any;
+  forgotForm: any;
+  registerForm: any;
   loginFlag = true;
   forgotFlag = false;
   newUser = true;
   usernameSuffix = false;
+  hide = false;
   registerMessage = 'To get complete access for content';
   loginMessage = 'To get latest updates/notifications';
   displayLogin = 'Already Registered User!';
   displayRegister = 'New User? Register Here';
-  passwordTooltip = 'Password must be combination of Capital Letter, Small Letter, Number and Special Characters[!@#$%&*]';
+  passwordTooltip = 'Password must have minimum 8 characters length & combination of Capital Letter/s, Small Letter/s, Number/s and Special Character/s [!@#$%&*]';
 
-  loginMessages = ValidationMessages.login_validation_messages;
-  forgotMessages = ValidationMessages.forgot_validation_messages;
+  errorMessages = ValidationMessages.validationErrorMessages;
 
   constructor(private httpClient: HttpClient, private formBuilder: FormBuilder, private matDialogRef: MatDialogRef<LoginRegisterComponent>) {
+    this.initializeForms();
+  }
+
+  ngOnInit() {
+    this.loginForm.get('username').valueChanges.subscribe((value: any) => {
+      value === '' ? this.usernameSuffix = false : isNaN(value) ? this.usernameSuffix = false : this.usernameSuffix = true;
+    });
+
+    this.registerForm.get('username').valueChanges.subscribe((value: any) => {
+      (value === '' || value === null || value === undefined) ? this.usernameSuffix = false : isNaN(value) ? this.usernameSuffix = false : this.usernameSuffix = true;
+    });
+  }
+
+  initializeForms() {
     this.loginForm = this.formBuilder.group({
       username: ['', Validators.compose([
         Validators.required, UsernameValidator.validUsername
       ])],
-      password: ['', Validators.required]
+      password: ['', Validators.compose([
+        Validators.required, Validators.minLength(8), Validators.pattern('^(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])(?=.*[!@#$%&*])[A-Za-z0-9!@#$%&*]{8,}$')
+      ])]
     });
 
     this.forgotForm = this.formBuilder.group({
@@ -48,17 +65,22 @@ export class LoginRegisterComponent implements OnInit {
         Validators.required, Validators.pattern('^[0-9]+$'), Validators.minLength(6)
       ])],
       password: ['', Validators.compose([
-        Validators.required, Validators.minLength(6), Validators.pattern('^[a-z][A-Z][0-9][!@#$%&*]+$')
+        Validators.required, Validators.minLength(8), Validators.pattern('^(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])(?=.*[!@#$%&*])[A-Za-z0-9!@#$%&*]{8,}$')
+      ])]
+    });
+
+    this.registerForm = this.formBuilder.group({
+      username: ['', Validators.compose([
+        Validators.required, UsernameValidator.validUsername
+      ])],
+      otp: ['', Validators.compose([
+        Validators.required, Validators.pattern('^[0-9]+$'), Validators.minLength(6)
+      ])],
+      password: ['', Validators.compose([
+        Validators.required, Validators.minLength(8), Validators.pattern('^(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])(?=.*[!@#$%&*])[A-Za-z0-9!@#$%&*]{8,}$')
       ])]
     });
   }
-
-  ngOnInit() {
-    this.loginForm.get('username').valueChanges.subscribe(value => {
-      value === '' ? this.usernameSuffix = false : isNaN(value) ? this.usernameSuffix = false : this.usernameSuffix = true;
-    });
-  }
-
 
   onSubmit(value) {
     //this.onClose();
@@ -69,8 +91,7 @@ export class LoginRegisterComponent implements OnInit {
   }
 
   login(value) {
-    console.log(value);
-    console.log(this.loginForm);
+
   }
 
   forgotPassword() {
@@ -79,6 +100,7 @@ export class LoginRegisterComponent implements OnInit {
       this.userFocus.nativeElement.focus();
     } else {
       this.forgotFlag = true;
+      this.hide = false;
       this.forgotForm.controls['username'].setValue(this.loginForm.get('username').value);
       this.forgotForm.controls['username'].disable();
     }
@@ -90,20 +112,30 @@ export class LoginRegisterComponent implements OnInit {
   }
 
   resetPassword(value) {
-    console.log(value);
-    console.log(this.forgotForm);
-  }
-
-  resendOtp(){
 
   }
 
-  validateRegisterUser(username) {
-    console.log(username);
+  resendOtp() {
+
   }
 
   register() {
 
+  }
+
+  generateOtp() {
+    this.newUser = false;
+  }
+
+  switch() {
+    this.loginFlag = !this.loginFlag;
+    this.forgotFlag = false;
+    this.newUser = true;
+    this.loginForm.reset();
+    this.usernameSuffix = false;
+    this.forgotForm.reset();
+    this.registerForm.reset();
+    this.hide = false;
   }
 
 }
