@@ -1,7 +1,7 @@
 import { AbstractControl, Validators, FormControl } from '@angular/forms';
 
 export class ValidationMessages {
-    public static login_validation_messages = {
+    public static validationErrorMessages = {
         'username': [
             { type: 'required', message: 'Email/Mobile Number Required' },
             { type: 'validUsernameEmail', message: 'Username must be valid Email' },
@@ -9,20 +9,13 @@ export class ValidationMessages {
         ],
         'password': [
             { type: 'required', message: 'Password Required' },
-            { type: 'pattern', message: 'Username must be Email/Mobile Number' }
-        ]
-    };
-
-    public static forgot_validation_messages = {
+            { type: 'pattern', message: 'Password as invalid pattern' },
+            { type: 'minlength', message: 'Password must have minimum 8 characters' }
+        ],
         'otp': [
             { type: 'required', message: 'OTP is required' },
             { type: 'pattern', message: 'OTP is always numerical' },
             { type: 'minlength', message: 'OTP must contain 6 digits' }
-        ],
-        'password': [
-            { type: 'required', message: 'Password Required' },
-            { type: 'pattern', message: 'Password as invalid pattern' },
-            { type: 'minlength', message: 'Password must have minimum 6 characters' }
         ]
     };
 }
@@ -31,12 +24,10 @@ export class ValidationMessages {
 
 export class UsernameValidator {
 
-
     static validUsername = (fc: FormControl) => {
         let emailPattern = '^[a-z0-9._%+-]+@[a-z0-9.-]+\\.[a-z]{2,4}$';
         let phonePattern = '^[0-9]+$';
-
-        if (fc.value !== '') {
+        if (fc.value !== '' && fc.value !== null && fc.value !== undefined) {
             if (isNaN(fc.value)) {
                 if (fc.value.match(emailPattern)) {
                     return null;
